+++ CON_469/CON_469_DFF/coolidea.py
@@ -1,7 +1,6 @@
 import os
 
 import pandas as pd
-import numpy as np
 import seaborn as sns
 import matplotlib.pyplot as plt
