+++ CON_469/CON_469_DFF/matrix3.py
@@ -1,6 +1,4 @@
-import os
 import pandas as pd
-import numpy as np
 import seaborn as sns
 import matplotlib.pyplot as plt
