+++ CON_469/CON_469_DFF/network.py
@@ -1,5 +1,4 @@
 import pandas as pd
-import numpy as np
 import networkx as nx
 import matplotlib.pyplot as plt
