+++ CON_469/CON_469_DFF/network2.py
@@ -1,5 +1,4 @@
 import pandas as pd
-import numpy as np
 import matplotlib.pyplot as plt
 import networkx as nx
