+++ GRO_823/GRO_823_DFF/1-3minutes.py
@@ -1,5 +1,4 @@
 import pandas as pd
-import numpy as np
 import seaborn as sns
 import matplotlib.pyplot as plt
 import os
