+++ GRO_823/GRO_823_DFF/Clustermap_variations_single_window.py
@@ -1,5 +1,4 @@
 import pandas as pd
-import numpy as np
 from sklearn.preprocessing import StandardScaler
 from sklearn.decomposition import PCA
 from sklearn.cluster import (
