+++ GRO_823/GRO_823_events/test/individual_correlational_matrices_with_data_values.py
@@ -1,6 +1,5 @@
 import pandas as pd
 import numpy as np
-import networkx as nx
 import matplotlib.pyplot as plt
 import seaborn as sns
 import os
