+++ HeatMaps.egg-info/PKG-INFO
@@ -1,0 +1,14 @@
+Metadata-Version: 2.1
+Name: HeatMaps
+Version: 0.1.0
+Summary: Description of my package
+Home-page: 
+Author: 
+Author-email: 
+Requires-Dist: pandas>=1.3.3
+Requires-Dist: numpy>=1.21.2
+Requires-Dist: seaborn>=0.11.2
+Requires-Dist: matplotlib>=3.4.3
+Requires-Dist: networkx>=2.6.3
+Requires-Dist: plotly>=5.3.1
+Requires-Dist: scipy>=1.7.1
