+++ HeatMaps.egg-info/requires.txt
@@ -1,0 +1,7 @@
+pandas>=1.3.3
+numpy>=1.21.2
+seaborn>=0.11.2
+matplotlib>=3.4.3
+networkx>=2.6.3
+plotly>=5.3.1
+scipy>=1.7.1
